move pulsingicon to individual module.
